devtools/server/tests/mochitest/test_inspector-insert.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <meta charset="utf-8">
  <title>Test for Bug </title>

  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
  <script type="application/javascript;version=1.8" src="inspector-helpers.js"></script>
  <script type="application/javascript;version=1.8">
const inspector = require("devtools/server/actors/inspector");

window.onload = function() {
  SimpleTest.waitForExplicitFinish();
  runNextTest();
}

var gWalker = null;
var gClient = null;

function assertOwnership() {
  return assertOwnershipTrees(gWalker);
}

addTest(function setup() {
  let url = document.getElementById("inspectorContent").href;
  attachURL(url, function(err, client, tab, doc) {
    gInspectee = doc;
    let {InspectorFront} = require("devtools/client/fronts/inspector");
    let inspector = InspectorFront(client, tab);
    promiseDone(inspector.getWalker().then(walker => {
      ok(walker, "getWalker() should return an actor.");
      gClient = client;
      gWalker = walker;
    }).then(runNextTest));
  });
});

addAsyncTest(function* testRearrange() {
  let longlist = yield gWalker.querySelector(gWalker.rootNode, "#longlist");
  let children = yield gWalker.children(longlist);
  let nodeA = children.nodes[0];
  is(nodeA.id, "a", "Got the expected node.");

  // Move nodeA to the end of the list.
  yield gWalker.insertBefore(nodeA, longlist, null);
  ok(!gInspectee.querySelector("#a").nextSibling, "a should now be at the end of the list.");
  children = yield gWalker.children(longlist);
  is(nodeA, children.nodes[children.nodes.length - 1], "a should now be the last returned child.");

  // Now move it to the middle of the list.
  let nextNode = children.nodes[13];
  yield gWalker.insertBefore(nodeA, longlist, nextNode);
  let sibling =
    new inspector._documentWalker(gInspectee.querySelector("#a"), window).nextSibling();
  is(sibling, nextNode.rawNode(), "Node should match the expected next node.");
  children = yield gWalker.children(longlist);
  is(nodeA, children.nodes[13], "a should be where we expect it.");
  is(nextNode, children.nodes[14], "next node should be where we expect it.");

  runNextTest();
});

addAsyncTest(function* testInsertInvalidInput() {
  let longlist = yield gWalker.querySelector(gWalker.rootNode, "#longlist");
  let children = yield gWalker.children(longlist);
  let nodeA = children.nodes[0];
  let nextSibling = children.nodes[1];

  // Now move it to the original location and make sure no mutation happens.
  let hasMutated = false;
  let observer = new gInspectee.defaultView.MutationObserver(() => {
    hasMutated = true;
  });
  observer.observe(longlist.rawNode(), {
    childList: true,
  });

  yield gWalker.insertBefore(nodeA, longlist, nodeA);
  ok(!hasMutated, "hasn't mutated");
  hasMutated = false;

  yield gWalker.insertBefore(nodeA, longlist, nextSibling);
  ok(!hasMutated, "still hasn't mutated after inserting before nextSibling");
  hasMutated = false;

  yield gWalker.insertBefore(nodeA, longlist);
  ok(hasMutated, "has mutated after inserting with null sibling");
  hasMutated = false;

  yield gWalker.insertBefore(nodeA, longlist);
  ok(!hasMutated, "hasn't mutated after inserting with null sibling again");

  observer.disconnect();
  runNextTest();
});

addTest(function cleanup() {
  delete gWalker;
  delete gClient;
  runNextTest();
});

  </script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=">Mozilla Bug </a>
<a id="inspectorContent" target="_blank" href="inspector-traversal-data.html">Test Document</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
</pre>
</body>
</html>
